renames to slices
