+++ src/components/modals/add-section-modal/add-modal-content.tsx
@@ -17,9 +17,9 @@ import { setValueFromEvent } from 'utils';
 import { useAppService } from 'contexts/app';
 import { useAddSection } from 'entities/section-entity';
 import { Section } from 'entities/types';
-import { onToggleAddModal } from 'reducers/modals';
+import { onToggleAddModal } from 'slices/modals';
 import { selectLoading, selectCreatedSection } from 'selectors';
-import { onResetCreatedSection } from 'reducers/created-section';
+import { onResetCreatedSection } from 'slices/created-section';
 import clsx from 'clsx';
 
 const AddSectionModalContent: React.FC<ModalContentProps> = ({ onClose }) => {
