+++ src/lib/filter-current-section-middleware.ts
@@ -1,5 +1,5 @@
 import { AppMiddleware } from 'global-types';
-import { onFetchForCloseSuccess, updateCloseSectionPayload } from 'reducers/close-section-payload';
+import { closeSectionDataReceived, updateCloseSectionPayload } from 'slices/close-section-payload';
 import { selectWithoutCurrentSection } from 'selectors/filter-sections';
 
 // removes the section to be closed from options
@@ -8,7 +8,7 @@ export const filterCurrentSectionMiddleware: AppMiddleware = ({
 }) => dispatch => action => {
     dispatch(action);
 
-    if (action.type === onFetchForCloseSuccess.type) {
+    if (action.type === closeSectionDataReceived.type) {
         const state = getState();
         const withoutCurrent = selectWithoutCurrentSection(state);
         dispatch(updateCloseSectionPayload(withoutCurrent));
