+++ src/pages/close-section/actions.ts
@@ -11,33 +11,28 @@ import {
 import { Dispatch } from 'redux';
 import { SectionsService } from 'services/section';
 import {
-    updateCloseSectionPayload,
-    onFetchForCloseSuccess,
-    onFetchFromStorageSuccess,
+    closeSectionDataReceived,
+    dataFromStorageReceived,
     onChangeInterventionSection,
     onUpdateInterventionIndicatorsState,
     onUpdateTravelSection,
     onUpdateActionPointSection,
     onUpdateTPMSections
-} from 'reducers/close-section-payload';
-import { onCurrentActiveSection } from 'reducers/current-active-section';
-import { requestStarted } from 'reducers/loading';
-import { onThrowError } from 'reducers/error';
-import { onSetModuleEditingName } from 'reducers/module-editing-name';
-import { onSuccessCloseSection } from 'reducers/closed-section-success';
+} from 'slices/close-section-payload';
+import { currentActiveSectionChanged } from 'slices/current-active-section';
+import { requestStarted } from 'slices/loading';
+import { onThrowError } from 'slices/error';
+import { onSetModuleEditingName } from 'slices/module-editing-name';
+import { onSuccessCloseSection } from 'slices/closed-section-success';
 import { getCloseSectionPrefixKey } from 'lib/sections';
 import { ActionBarKeys } from './types';
-import { setCloseSectionActionBar, onSetViewCloseSummary } from 'reducers/ui';
-
-export const onResetCloseSectionPayload = (dispatch: Dispatch) => {
-    dispatch(updateCloseSectionPayload(null));
-};
+import { setCloseSectionActionBar, onSetViewCloseSummary } from 'slices/ui';
 
 export const onFetchDataCloseSection = (
     services: { backendService: BackendService; storageService: StorageService },
     payload: FetchStoragePayload
 ) => async (dispatch: Dispatch) => {
-    dispatch(onCurrentActiveSection(Number(payload.id)));
+    dispatch(currentActiveSectionChanged(Number(payload.id)));
 
     const { backendService, storageService } = services;
     const key = getCloseSectionPrefixKey(payload);
@@ -54,9 +49,9 @@ export const onFetchDataCloseSection = (
             return;
         }
 
-        dispatch(onFetchForCloseSuccess(dataFromServer));
+        dispatch(closeSectionDataReceived(dataFromServer));
     } else {
-        dispatch(onFetchFromStorageSuccess(dataFromStorage));
+        dispatch(dataFromStorageReceived(dataFromStorage));
     }
 };
