+++ src/pages/close-section/edit-wrapper.tsx
@@ -4,7 +4,7 @@ import Box from 'components/box';
 import { IconButton, makeStyles, createStyles, Theme, Paper, Typography, Container } from '@material-ui/core';
 import { useDispatch } from 'react-redux';
 import { useIconButtonStyles } from 'components/table/styles';
-import { onSetModuleEditingName } from 'reducers/module-editing-name';
+import { onSetModuleEditingName } from 'slices/module-editing-name';
 
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
