+++ src/slices/__tests__/sections-list.spec.ts
@@ -1,4 +1,4 @@
-import { onGetSectionsSuccess } from 'reducers/sections';
+import { onGetSectionsSuccess } from '../sections';
 import { testSectionsList } from './fixtures/sections';
 import configureAppStore from 'lib/create-store';
 import { selectSections, selectAllSections } from 'selectors';
@@ -10,13 +10,13 @@ describe('Sections list ', () => {
         store = configureAppStore();
     });
 
-    test('should add sections to store', async () => {
+    test('sections received from api call are added to store correctly', async () => {
         await store.dispatch(onGetSectionsSuccess(testSectionsList));
         const storeSections = selectAllSections(store.getState());
         expect(storeSections).toEqual(testSectionsList);
     });
 
-    test('should only show inactive sections in list by default', async () => {
+    test('default selector for sections should only include inactive sections in list by default', async () => {
         await store.dispatch(onGetSectionsSuccess(testSectionsList));
         const storeSections = selectSections(store.getState());
         expect(storeSections).not.toContain(testSectionsList[2]);
