+++ src/slices/close-section-payload.ts
@@ -13,8 +13,8 @@ export const closeSectionPayload = createSlice({
     name: 'closeSectionPayload',
     initialState,
     reducers: {
-        onFetchFromStorageSuccess: (state, action) => action.payload,
-        onFetchForCloseSuccess: (state, action) => action.payload,
+        dataFromStorageReceived: (state, action) => action.payload,
+        closeSectionDataReceived: (state, action) => action.payload,
         updateCloseSectionPayload: (state, action) => action.payload,
         onChangeInterventionSection: (state, action) => {
             const { id, sections } = action.payload;
@@ -43,9 +43,9 @@ export const closeSectionPayload = createSlice({
 });
 
 export const {
-    onFetchFromStorageSuccess,
+    dataFromStorageReceived,
     updateCloseSectionPayload,
-    onFetchForCloseSuccess,
+    closeSectionDataReceived,
     onUpdateTravelSection,
     onChangeInterventionSection,
     onUpdateActionPointSection,
