+++ src/slices/current-active-section.ts
@@ -5,11 +5,11 @@ const currentActiveSectionId = createSlice({
     name: 'currentActiveSectionId',
     initialState: -1,
     reducers: {
-        onCurrentActiveSection: (state, action) => action.payload
+        currentActiveSectionChanged: (state, action) => action.payload
     },
     extraReducers: {
         [renderSectionsList.type]: () => -1
     }
 });
-export const { onCurrentActiveSection } = currentActiveSectionId.actions;
+export const { currentActiveSectionChanged } = currentActiveSectionId.actions;
 export const { reducer: cuurentActiveSectionReducer } = currentActiveSectionId;
