+++ src/slices/index.ts
@@ -1,14 +1,14 @@
 import { combineReducers } from 'redux';
-import { closeSectionPayloadReducer } from 'reducers/close-section-payload';
-import { sectionsReducer } from 'reducers/sections';
-import { createdSectionReducer } from 'reducers/created-section';
-import { mergedSectionReducer } from 'reducers/merged-section';
-import { errorReducer } from 'reducers/error';
-import { loadingReducer } from 'reducers/loading';
-import { moduleEditingNameReducer } from 'reducers/module-editing-name';
-import { cuurentActiveSectionReducer } from 'reducers/current-active-section';
-import { closedSectionSuccessReducer } from 'reducers/closed-section-success';
-import { uiReducer } from 'reducers/ui';
+import { closeSectionPayloadReducer } from 'slices/close-section-payload';
+import { sectionsReducer } from 'slices/sections';
+import { createdSectionReducer } from 'slices/created-section';
+import { mergedSectionReducer } from 'slices/merged-section';
+import { errorReducer } from 'slices/error';
+import { loadingReducer } from 'slices/loading';
+import { moduleEditingNameReducer } from 'slices/module-editing-name';
+import { cuurentActiveSectionReducer } from 'slices/current-active-section';
+import { closedSectionSuccessReducer } from 'slices/closed-section-success';
+import { uiReducer } from 'slices/ui';
 import { userReducer } from './user';
 import { namesFromSplitReducer } from './names-from-split';
 import { inProgressItemsReducer } from './in-progress-items';
