+++ test/test-utils.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import { Provider } from 'react-redux';
+
 import { render, RenderOptions } from '@testing-library/react';
 import AppProviders from '../src/components/app-providers';
 
@@ -15,3 +17,13 @@ export * from '@testing-library/react';
 
 // override render method
 export { connectedRender };
+
+export function renderWithRedux(ui, { store }) {
+    return {
+        ...render(<Provider store={store}>{ui}</Provider>),
+        // adding `store` to the returned utilities to allow us
+        // to reference it in our tests (just try to avoid using
+        // this to test implementation details).
+        store
+    };
+}
